Update download URL and naming
The naming and URL changed from `bittorrent-sync` to `resilio-sync`. So maybe `bittorrent-sync` cask should point to a newly added `resilio-sync` cask? Not sure how you usually handle this case. I'm happy to improve this in any way. Just wanted to start a process here. I can also close this and open a separate issue for this.